time: improve printing of timespecs
And cast types down to ulong, where possible. This causes the compiler
to work on 32 bit types instead of 64 bit types. Here in particular, we
have divisions that allows to emit the compiler 32-bit variants.

+++ include/grinch/time.h
@@ -21,7 +21,8 @@
 #define MS_TO_NS(x)	(US_TO_NS((x) * 1000L))
 #define S_TO_NS(x)	(MS_TO_NS((x) * 1000L))
 
-#define NS_TO_MS(x)	((x) / MS_TO_NS(1))
+/* We know that the divisor is definitely 32 bit, so cast it to ulong */
+#define NS_TO_MS(x)	((x) / (unsigned long)MS_TO_NS(1))
 
 #define NSEC_PER_SEC	S_TO_NS(1)
 #define MSEC_PER_SEC	MS_TO_NS(1)
@@ -34,8 +35,12 @@
 
 #define HZ_TO_NS(x)	(NSEC_PER_SEC / (x))
 
-#define PR_TS_FMT	"%04llu.%03llu"
-#define PR_TS_PARAMS(x)	(x)->tv_sec, NS_TO_MS((x)->tv_nsec)
+/*
+ * Makes the assumption that the timespec is normalised. Hence we can cast
+ * tv_nsec to ulong.
+ */
+#define PR_TS_FMT	"%04llu.%03lu"
+#define PR_TS_PARAMS(x)	(x)->tv_sec, NS_TO_MS((unsigned long)(x)->tv_nsec)
 
 void ns_to_ts(time_t ns, struct timespec *ts);
 static inline void us_to_ts(time_t ns, struct timespec *ts)
